Remove Matlibplot test rendering, split tests with descripts, add BaseModel() tests
