+++ pybop/identification/observations.py
@@ -1,5 +1,6 @@
 import pybamm
 
+
 class Observed:
     """
     Class for experimental Observations.
