+++ pybop/identification/parameter_set.py
@@ -1,5 +1,6 @@
 import pybamm
 
+
 class ParameterSet:
     """
     Class for creating parameter sets in pybop.
@@ -7,6 +8,6 @@ class ParameterSet:
 
     def __new__(cls, method, name):
         if method.casefold() == "pybamm":
-                return pybamm.ParameterValues(name).copy()
+            return pybamm.ParameterValues(name).copy()
         else:
             raise ValueError("Only PybaMM parameter sets are currently implemented")
